Рефакторинг, исправление ошибок
ToDo и список багов перенесены в трелло.
Проект переведён на стандарт C++17.
Замена устаревших конструкций.
Повышение качества и читаемости кода.
Обновлён GNU Bison с 3.1 до 3.3.2.
Целое теперь занимает 8 байт.
Длинные целые теперь интерпретируются как 8 байтовое число с плавающей точкой.
Программы теперь могут содержать больше 32767 строк.
Размер представления данных теперь 8 байт, а не 24.
Рефакторинг структуры для хранения типовой информации и методов работы с ней.
Исправлена ошибка, приводившая к аварийному завершению работы на ‭2^31+1‬ попытке взять элемент из пустой очереди вычислителя.
Добавлена обработка ошибок при выполнении программ.
Добавлен аргумент запуска, задающий начальный кортеж.

+++ Parser/Ident.h
@@ -10,8 +10,8 @@ namespace Parser {
 	// Указатель на запись в таблице имен.
 	struct Ident
 	{
-		short Col;
-		short Line;
+		size_t Col;
+		size_t Line;
 		const std::string * Ptr;
 
 		bool operator==( const Ident & aOther ) const { return Ptr == aOther.Ptr; }
